Rename and reorganize commands
