+++ lib/fauxpaas/commands/rollback.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+require "fauxpaas"
+require "fauxpaas/commands/command"
+
+module Fauxpaas
+  module Commands
+
+    # Rollback to a previous cache
+    class Rollback < Command
+    end
+
+  end
+end
