+++ lib/fauxpaas/commands/syslog_follow.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+require "fauxpaas"
+require "fauxpaas/commands/command"
+
+module Fauxpaas
+  module Commands
+
+    # tail -f the system logs
+    class SyslogFollow < Command
+      def action
+        :syslog_follow
+      end
+
+      def execute
+        instance.interrogator.syslog_follow
+      end
+    end
+
+  end
+end
